refactoring pytest fixtures
